Added query validation utility
