+++ tpch-dbgen/answers/q3.out
@@ -1,4 +1,4 @@
-l_orderkey            |revenue                                  |o_orderdat|o_shippriority      
+l_orderkey            |revenue                                  |o_orderdate|o_shippriority
                2456423|406181.01|1995-03-05|                   0
                3459808|405838.70|1995-03-04|                   0
                 492164|390324.06|1995-02-19|                   0
